requirements.txt
Flask
flask_cors
Werkzeug
openai
python-dotenv
requests
gunicorn
tweepy